refactor: TotalPoints 컴포넌트 css 분리
