+++ src/components/FocusComponents/TotalPoints/TotalPoints.module.css
@@ -1,0 +1,10 @@
+.pointContainer {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-top: 16px;
+}
+.pointContainer > p {
+  font-size: 18px;
+  color: var(--gray_818181);
+}
