+++ src/pages/TodayFocusPage/TodayFocus.jsx
@@ -1,7 +1,7 @@
 import styles from './TodayFocus.module.css';
 import { useEffect, useRef, useState } from 'react';
 import FocusHeader from '../../components/FocusComponents/FocusHeader/FocusHeader';
-import TotalPoints from '../../components/FocusComponents/TotalPoints';
+import TotalPoints from '../../components/FocusComponents/TotalPoints/TotalPoints';
 import TargetDuration from '../../components/FocusComponents/TargetDuration/TargetDuration';
 import Timer from '../../components/FocusComponents/Timer/Timer';
 import { useParams } from 'react-router-dom';
